lib.image: leverage perp rc.log
Use perp logger to send process stdout/stderr to fluentd.

For any service that wants to send stdout/stderr to fluentd we
create a rc.log file. When this file is present perp will pass
stdout/stderr of the process to the logger specified in the rc.log
file.

In rc.log we create the named pipe and start tee on this pipe. We
actually don't need to wait on fluentd or fluentd doesn't need to
have the pipe when it starts.

Also, add test for stderr output.